Modifiers: Rename "Empty Modifier" to "Geometry Nodes"
Reasons to keep calling it Geometry Nodes:
* People can search for Geometry or Nodes and find it.
* It makes a connection with the editor called Geometry Nodes.

Interesting enough these were similar reasons to the original discussion
we had when Geometry Nodes first got added to Blender. Basically the
reasoning is still valid.

Counter-arguments:
* Over-time more and more modifiers will be geometry-nodes based, it
  could be seem as strange to single out one of them.
* Once we have different node-base systems that may also work as
  modifiers (e.g., collection modifiers?) we will need a more unified
  name.

The response to the latter is that once we get to this we will also
rename the editor. Until then we keep what is working.